Made the `/transfer` and `/rules` commands publicly available
