+++ bot/handlers/cogs/private.py
@@ -2,16 +2,13 @@ from datetime import timedelta, datetime
 from statistics import mean
 
 import disnake
-from disnake import AppCommandInter, AllowedMentions, Webhook, Embed, Localized
+from disnake import AppCommandInter, Embed, Localized
 from disnake.ext.commands import Cog, slash_command, Param
 from loguru import logger
 
 from bot.data.clubpenguin.penguin import Login, Penguin
 from bot.data.clubpenguin.transactions import Transactions
-from bot.handlers.button import Rules
 from bot.misc.constants import (
-    embedRuleImageRu,
-    embedRuleRu,
     embedAboutImage,
     embedAbout,
     guild_ids,
@@ -24,46 +21,6 @@ class PrivateCommands(Cog):
         self.bot = bot
 
         logger.info(f"Loaded {len(self.get_application_commands())} private app commands")
-
-    @slash_command(guild_ids=guild_ids)
-    async def transfer(self, inter: AppCommandInter, forum: disnake.ForumChannel):
-        """
-        Transfer images from the current channel to the forum {{TRANSFER}}
-
-        Parameters
-        ----------
-        inter: AppCommandInter
-        forum: disnake.ForumChannel
-            Target forum channel on the current server {{FORUM}}
-        """
-        await inter.response.defer()
-        source_channel = inter.channel
-        destination_channel = disnake.utils.get(inter.guild.channels, id=forum.id)
-
-        async for message in source_channel.history(limit=None):
-            if message.attachments:
-                content = f", comment: {message.content}" if message.content else ''
-                await destination_channel.create_thread(name=f"{message.author.display_name}",
-                                                        content=f'Author: {message.author.mention}{content}',
-                                                        files=[await attachment.to_file() for attachment in
-                                                               message.attachments],
-                                                        allowed_mentions=AllowedMentions(users=False)
-                                                        )
-        await inter.edit_original_response(f"Success transferred in {forum.mention}!")
-
-    @slash_command(guild_ids=guild_ids)
-    async def rules(self, inter: AppCommandInter):
-        """
-        Send rules embed {{RULES}}
-
-        Parameters
-        ----------
-        inter: AppCommandInter
-        """
-        webhook: Webhook = await inter.channel.create_webhook(name="CPPS.APP", avatar=avatarImageBytearray)
-        message = await webhook.send(embeds=[embedRuleImageRu, embedRuleRu], wait=True)
-        await message.edit(view=Rules(message))
-        await inter.send("Success", ephemeral=True)
 
     @slash_command(guild_ids=guild_ids)
     async def about(self, inter: AppCommandInter):
